I2cCtrl add address filter
